Updated dependencies and upgraded to Scala.js 1.0.0-RC2

+++ core/js/src/main/scala/perfolation/JavaScriptCrossDate.scala
@@ -5,18 +5,18 @@ import scala.scalajs.js.Date
 
 class JavaScriptCrossDate(val date: Date) extends AnyVal with CrossDate {
   override def milliseconds: Long = date.getTime().toLong
-  override def hour24: Int = date.getHours()
-  override def minuteOfHour: Int = date.getMinutes()
-  override def secondOfMinute: Int = date.getSeconds()
-  override def milliOfSecond: Int = date.getMilliseconds()
+  override def hour24: Int = date.getHours().toInt
+  override def minuteOfHour: Int = date.getMinutes().toInt
+  override def secondOfMinute: Int = date.getSeconds().toInt
+  override def milliOfSecond: Int = date.getMilliseconds().toInt
   override def isAM: Boolean = hour24 < 12
-  override def timeZoneOffsetMillis: Int = -date.getTimezoneOffset() * 60 * 1000
-  override def year: Int = date.getFullYear()
-  override def month: Int = date.getMonth()
-  override def dayOfWeek: Int = date.getDay() + 1
-  override def dayOfMonth: Int = date.getDate()
+  override def timeZoneOffsetMillis: Int = (-date.getTimezoneOffset() * 60 * 1000).toInt
+  override def year: Int = date.getFullYear().toInt
+  override def month: Int = date.getMonth().toInt
+  override def dayOfWeek: Int = date.getDay().toInt + 1
+  override def dayOfMonth: Int = date.getDate().toInt
   override def dayOfYear: Int = {
-    val beginning = new Date(date.getFullYear(), 0, 0)
+    val beginning = new Date(date.getFullYear().toInt, 0, 0)
     val diff = date.getTime() - beginning.getTime()
     val oneDay = 1000.0 * 60.0 * 60.0 * 24.0
     math.floor(diff / oneDay).toInt - 1
