+++ core/js/src/main/scala/perfolation/Platform.scala
@@ -3,7 +3,7 @@ package perfolation
 import scala.scalajs.js.Date
 
 object Platform {
-  val startOfYear: Date = new Date(new Date().getFullYear(), 0, 0)
+  val startOfYear: Date = new Date(new Date().getFullYear().toInt, 0, 0)
 
   def createDate(l: Long): CrossDate = new JavaScriptCrossDate(new Date(l.toDouble))
 }
